realized TaskTimer.java was in the unintended package, moved to correct .util
