+++ src/net/sf/memoranda/ui/TaskDialog.java
@@ -40,8 +40,8 @@ import javax.swing.JCheckBox;
 
 import net.sf.memoranda.CurrentProject;
 import net.sf.memoranda.date.CalendarDate;
-import net.sf.memoranda.ui.htmleditor.util.TaskTimer;
 import net.sf.memoranda.util.Local;
+import net.sf.memoranda.util.TaskTimer;
 
 import javax.swing.JToggleButton;
